Adding the ChrootManager class
Chroot manager class can be used to setup
a chroot enviroment including an arbitrary
list of bind mounts. The provided paths to bind
are mounted in the chroot to the same actual root
path they are bound from.

This class is useful to setup a chroot envirment
based on the root-tree created on prepare step.

Signed-off-by: David Cassany <dcassany@suse.com>